requirements.txt
sseclient-py=1.8.0
urllib3=2.0.4